LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/browsers/history/the-history-interface/history_back_cross_realm_method-expected.txt



Harness Error (TIMEOUT), message = null

PASS history.back() uses this's associated document's browsing context to determine if navigation is allowed
